+++ internal/model/model.go
@@ -92,6 +92,8 @@ type (
 		TransactionID *uuid.UUID  `json:"transaction_id,omitempty"`
 		UserID        string      `json:"user_id"`
 		IsPaid        bool        `json:"is_paid"`
+		IsRecurrent   bool        `json:"is_recurrent"`
+		RecurrentID   *uuid.UUID  `json:"recurrent_id"`
 		StatusID      int         `json:"status_id,omitempty"`
 		WalletID      *uuid.UUID  `json:"wallet_id,omitempty"`
 		Wallet        Wallet      `json:"wallets,omitempty"`
@@ -106,6 +108,19 @@ type (
 	}
 
 	MovementList []Movement
+
+	RecurrentMovement struct {
+		ID            *uuid.UUID `json:"id,omitempty" gorm:"primaryKey"`
+		Description   string     `json:"description,omitempty"`
+		Amount        float64    `json:"amount"`
+		InitialDate   *time.Time `json:"initial_date"`
+		EndDate       *time.Time `json:"end_date"`
+		UserID        string     `json:"user_id"`
+		CategoryID    *uuid.UUID `json:"category_id,omitempty"`
+		SubCategoryID *uuid.UUID `json:"sub_category_id,omitempty"`
+		WalletID      *uuid.UUID `json:"wallet_id,omitempty"`
+		TypePaymentID int        `json:"type_payment_id,omitempty"`
+	}
 
 	Transaction struct {
 		TransactionID *uuid.UUID     `json:"transaction_id"`
@@ -257,4 +272,17 @@ func (pt *Transaction) Consolidate() {
 
 func (b *Balance) Consolidate() {
 	b.PeriodBalance = b.Income + b.Expense
+}
+
+func ToRecurrentMovement(movement Movement) RecurrentMovement {
+	return RecurrentMovement{
+		Description:   movement.Description,
+		Amount:        movement.Amount,
+		InitialDate:   movement.Date,
+		UserID:        movement.UserID,
+		CategoryID:    movement.CategoryID,
+		SubCategoryID: movement.SubCategoryID,
+		WalletID:      movement.WalletID,
+		TypePaymentID: movement.TypePaymentID,
+	}
 }