+++ CHANGELOG.md
@@ -8,6 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- Created model and Add recurrent movement [PR#55](https://github.com/silvioubaldino/personal-finance/pull/55)
 - Added update wallet when deleting movement [PR#54](https://github.com/silvioubaldino/personal-finance/pull/54)
 
 ## Release - v1.0.1 - 09-11-2024